+++ sdk/honcho/sync_client.py
@@ -1,16 +1,43 @@
 import json
+import uuid
 from typing import Dict, Optional 
 import httpx
-from .schemas import Message
+from .schemas import Message, Metamessage
 
-class GetSessionResponse:
-    def __init__(self, client, response: Dict):
-        self.client = client
+class GetPage:
+    """Base class for receiving Paginated API results"""
+    def __init__(self, response: Dict) -> None:
+        """Constructor for Page with relevant information about the results and pages
+
+        Args:
+            response (Dict): Response from API with pagination information
+        """
         self.total = response["total"]
         self.page = response["page"]
         self.page_size = response["size"]
         self.pages = response["pages"]
-        self.sessions = [
+        self.items =[]
+
+    def next(self):
+        """Shortcut method to Get the next page of results"""
+        pass
+
+class GetSessionPage(GetPage):
+    """Paginated Results for Get Session Requests"""
+
+    def __init__(self, client, options: Dict, response: Dict):
+        """Constructor for Page Result from Session Get Request
+        
+        Args:
+            client (Client): Honcho Client
+            options (Dict): Options for the request used mainly for next() to filter queries. The two parameters available are user_id which is required and location_id which is optional
+            response (Dict): Response from API with pagination information
+        """
+        super().__init__(response)
+        self.client = client
+        self.user_id = options["user_id"]
+        self.location_id = options["location_id"]
+        self.items = [
             Session(
                 client=client,
                 id=session["id"],
@@ -23,36 +50,85 @@ class GetSessionResponse:
         ]
        
     def next(self):
+        """Get the next page of results
+        Returns:
+            GetSessionPage | None: Next Page of Results or None if there are no more sessions to retreive from a query
+        """
         if self.page >= self.pages:
             return None
-        user_id = self.sessions[0].user_id
-        location_id = self.sessions[0].location_id
-        return self.client.get_sessions(user_id, location_id, self.page + 1, self.page_size)
+        return self.client.get_sessions(self.user_id, self.location_id, self.page + 1, self.page_size)
 
-class GetMessageResponse:
+class GetMessagePage(GetPage):
+    """Paginated Results for Get Session Requests"""
+
     def __init__(self, session, response: Dict):
+        """Constructor for Page Result from Session Get Request
+        
+        Args:
+            session (Session): Session the returned messages are associated with
+            response (Dict): Response from API with pagination information
+        """
+        super().__init__(response)
         self.session = session
-        self.total = response["total"]
-        self.page = response["page"]
-        self.page_size = response["size"]
-        self.pages = response["pages"]
-        self.messages = [
+        self.items = [
                 Message(
                 session_id=session.id,
                 id=message["id"],
                 is_user=message["is_user"],
                 content=message["content"],
+                created_at=message["created_at"],
             )
             for message in response["items"]
         ]
 
     def next(self):
+        """Get the next page of results
+        Returns:
+            GetMessagePage | None: Next Page of Results or None if there are no more messages to retreive from a query
+        """
         if self.page >= self.pages:
             return None
         return self.session.get_messages((self.page + 1), self.page_size)
 
+class GetMetamessagePage(GetPage):
+    
+    def __init__(self, session, options: Dict, response: Dict) -> None:
+        """Constructor for Page Result from Metamessage Get Request
+        
+        Args:
+            session (Session): Session the returned messages are associated with
+            options (Dict): Options for the request used mainly for next() to filter queries. The two parameters available are message_id and metamessage_type which are both required
+            response (Dict): Response from API with pagination information
+        """
+        super().__init__(response)
+        self.session = session
+        self.message_id = options["message_id"]
+        self.metamessage_type = options["metamessage_type"]
+        self.items = [
+                Metamessage(
+                    id=metamessage["id"],
+                    message_id=metamessage["message_id"],
+                    metamessage_type=metamessage["metamessage_type"],
+                    content=metamessage["content"],
+                    created_at=metamessage["created_at"],
+                    )
+                for metamessage in response["items"]
+        ]
+
+    def next(self):
+        """Get the next page of results
+        Returns:
+            GetMetamessagePage | None: Next Page of Results or None if there are no more metamessages to retreive from a query
+        """
+        if self.page >= self.pages:
+            return None
+        return self.session.get_metamessages(metamessage_type=self.metamessage_type, message=self.message_id, page=(self.page + 1), page_size=self.page_size)
+
+
 
 class Client:
+    """Honcho API Client Object"""
+
     def __init__(self, app_id: str, base_url: str = "https://demo.honcho.dev"):
         """Constructor for Client"""
         self.base_url = base_url  # Base URL for the instance of the Honcho API
@@ -61,21 +137,23 @@ class Client:
 
     @property
     def common_prefix(self):
+        """Shorcut for common API prefix. made a property to prevent tampering"""
         return f"{self.base_url}/apps/{self.app_id}"
 
-    def get_session(self, user_id: str, session_id: int):
+    def get_session(self, user_id: str, session_id: uuid.UUID):
         """Get a specific session for a user by ID
 
         Args:
             user_id (str): The User ID representing the user, managed by the user
-            session_id (int): The ID of the Session to retrieve
+            session_id (uuid.UUID): The ID of the Session to retrieve
 
         Returns:
-            Dict: The Session object of the requested Session
+            Session: The Session object of the requested Session
 
         """
         url = f"{self.common_prefix}/users/{user_id}/sessions/{session_id}"
         response = self.client.get(url)
+        response.raise_for_status()
         data = response.json()
         return Session(
             client=self,
@@ -87,14 +165,16 @@ class Client:
         )
 
     def get_sessions(self, user_id: str, location_id: Optional[str] = None, page: int = 1, page_size: int = 50):
-        """Return sessions associated with a user
+        """Return sessions associated with a user paginated
 
         Args:
             user_id (str): The User ID representing the user, managed by the user
             location_id (str, optional): Optional Location ID representing the location of a session
+            page (int, optional): The page of results to return
+            page_size (int, optional): The number of results to return
 
         Returns:
-            list[Dict]: List of Session objects
+            GetSessionPage: Page or results for get_sessions query
 
         """
         url = f"{self.common_prefix}/users/{user_id}/sessions?page={page}&size={page_size}" + (
@@ -103,15 +183,29 @@ class Client:
         response = self.client.get(url)
         response.raise_for_status()
         data = response.json()
-        return GetSessionResponse(self, data)
+        options = {
+                "location_id": location_id,
+                "user_id": user_id
+                }
+        return GetSessionPage(self, options, data)
 
     def get_sessions_generator(self, user_id: str, location_id: Optional[str] = None):
+        """Shortcut Generator for get_sessions. Generator to iterate through all sessions for a user in an app
+
+        Args:
+            user_id (str): The User ID representing the user, managed by the user
+            location_id (str, optional): Optional Location ID representing the location of a session
+
+        Yields:
+            Session: The Session object of the requested Session
+
+        """
         page = 1
         page_size = 50
         get_session_response = self.get_sessions(user_id, location_id, page, page_size)
         while True:
             # get_session_response = self.get_sessions(user_id, location_id, page, page_size)
-            for session in get_session_response.sessions:
+            for session in get_session_response.items:
                 yield session
 
             new_sessions = get_session_response.next()
@@ -131,12 +225,13 @@ class Client:
             session_data (Dict, optional): Optional session metadata
 
         Returns:
-            Dict: The Session object of the new Session`
+            Session: The Session object of the new Session
 
         """
         data = {"location_id": location_id, "session_data": session_data}
         url = f"{self.common_prefix}/users/{user_id}/sessions"
         response = self.client.post(url, json=data)
+        response.raise_for_status()
         data = response.json()
         return Session(
             self,
@@ -149,10 +244,12 @@ class Client:
 
 
 class Session:
+    """Represents a single session for a user in an app"""
+
     def __init__(
         self,
         client: Client,
-        id: int,
+        id: uuid.UUID,
         user_id: str,
         location_id: str,
         session_data: dict | str,
@@ -172,13 +269,16 @@ class Session:
 
     @property
     def common_prefix(self):
+        """Shortcut for common API prefix. made a property to prevent tampering"""
         return f"{self.base_url}/apps/{self.app_id}"
 
     def __str__(self):
+        """String representation of Session"""
         return f"Session(id={self.id}, app_id={self.app_id}, user_id={self.user_id}, location_id={self.location_id}, session_data={self.session_data}, is_active={self.is_active})"
 
     @property
     def is_active(self):
+        """Returns whether the session is active - made property to prevent tampering"""
         return self._is_active
 
     def create_message(self, is_user: bool, content: str):
@@ -189,7 +289,7 @@ class Session:
             content (str): The content of the message
 
         Returns:
-            Dict: The Message object of the added message
+            Message: The Message object of the added message
 
         """
         if not self.is_active:
@@ -197,10 +297,103 @@ class Session:
         data = {"is_user": is_user, "content": content}
         url = f"{self.common_prefix}/users/{self.user_id}/sessions/{self.id}/messages"
         response = self.client.post(url, json=data)
+        response.raise_for_status()
         data = response.json()
-        return Message(session_id=self.id, id=data["id"], is_user=is_user, content=content)
+        return Message(session_id=self.id, id=data["id"], is_user=is_user, content=content, created_at=data["created_at"])
 
-    def get_messages(self, page: int = 1, page_size: int = 50) -> GetMessageResponse:
+    def get_message(self, message_id: uuid.UUID) -> Message:
+        """Get a specific message for a session based on ID
+
+        Args:
+            message_id (uuid.UUID): The ID of the Message to retrieve
+
+        Returns:
+            Message: The Message object
+
+        """
+        url = f"{self.common_prefix}/users/{self.user_id}/sessions/{self.id}/messages/{message_id}"
+        response = self.client.get(url)
+        response.raise_for_status()
+        data = response.json()
+        return Message(session_id=self.id, id=data["id"], is_user=data["is_user"], content=data["content"], created_at=data["created_at"])
+
+    def get_messages(self, page: int = 1, page_size: int = 50) -> GetMessagePage:
+        """Get all messages for a session
+
+        Args:
+            page (int, optional): The page of results to return
+            page_size (int, optional): The number of results to return per page
+
+        Returns:
+            GetMessagePage: Page of Message objects
+
+        """
+        url = f"{self.common_prefix}/users/{self.user_id}/sessions/{self.id}/messages?page={page}&size={page_size}"
+        response = self.client.get(url)
+        response.raise_for_status()
+        data = response.json()
+        return GetMessagePage(self, data)
+        
+    def get_messages_generator(self):
+        """Shortcut Generator for get_messages. Generator to iterate through all messages for a session in an app
+
+        Yields:
+            Message: The Message object of the next Message
+
+        """
+        page = 1
+        page_size = 50
+        get_messages_page= self.get_messages(page, page_size)
+        while True:
+            # get_session_response = self.get_sessions(user_id, location_id, page, page_size)
+            for message in get_messages_page.items:
+                yield message
+
+            new_messages = get_messages_page.next()
+            if not new_messages:
+                break
+           
+            get_messages_page = new_messages
+
+    def create_metamessage(self, message: Message, metamessage_type: str, content: str):
+        """Adds a metamessage to a session and links it to a specific message
+
+        Args:
+            message (Message): A message to associate the metamessage with
+            metamessage_type (str): The type of the metamessage arbitrary itentifier
+            content (str): The content of the metamessage
+
+        Returns:
+            Metamessage: The Metamessage object of the added metamessage
+
+        """
+        if not self.is_active:
+            raise Exception("Session is inactive")
+        data = {"metamessage_type": metamessage_type, "content": content, "message_id": message.id}
+        url = f"{self.common_prefix}/users/{self.user_id}/sessions/{self.id}/metamessages"
+        response = self.client.post(url, json=data)
+        response.raise_for_status()
+        data = response.json()
+        return Metamessage(id=data["id"], message_id=message.id, metamessage_type=metamessage_type, content=content, created_at=data["created_at"])
+
+
+    def get_metamessage(self, metamessage_id: uuid.UUID) -> Metamessage:
+        """Get a specific metamessage
+
+        Args:
+            message_id (uuid.UUID): The ID of the Message to retrieve
+
+        Returns:
+            Message: The Message object
+
+        """
+        url = f"{self.common_prefix}/users/{self.user_id}/sessions/{self.id}/metamessages/{metamessage_id}"
+        response = self.client.get(url)
+        response.raise_for_status()
+        data = response.json()
+        return Metamessage(id=data["id"], message_id=data["message_id"], metamessage_type=data["metamessage_type"], content=data["content"], created_at=data["created_at"])
+
+    def get_metamessages(self, metamessage_type: Optional[str] = None, message: Optional[Message] = None, page: int = 1, page_size: int = 50) -> GetMetamessagePage:
         """Get all messages for a session
 
         Args:
@@ -211,40 +404,51 @@ class Session:
             list[Dict]: List of Message objects
 
         """
-        url = f"{self.common_prefix}/users/{self.user_id}/sessions/{self.id}/messages?page={page}&size={page_size}"
+        url = f"{self.common_prefix}/users/{self.user_id}/sessions/{self.id}/metamessages?page={page}&size={page_size}"
+        if metamessage_type:
+            url += f"&metamessage_type={metamessage_type}"
+        if message:
+            url += f"&message_id={message.id}"
         response = self.client.get(url)
         response.raise_for_status()
         data = response.json()
-        return GetMessageResponse(self, data)
-        # return [
-        #     Message(
-        #         self,
-        #         id=message["id"],
-        #         is_user=message["is_user"],
-        #         content=message["content"],
-        #     )
-        #     for message in data
-        # ]
-    def get_messages_generator(self):
+        options = {
+                "metamessage_type": metamessage_type,
+                "message_id": message.id if message else None
+                }
+        return GetMetamessagePage(self, options, data)
+        
+    def get_metamessages_generator(self, metamessage_type: Optional[str] = None, message: Optional[Message] = None):
+        """Shortcut Generator for get_metamessages. Generator to iterate through all metamessages for a session in an app
+
+        Args:
+            metamessage_type (str, optional): Optional Metamessage type to filter by
+            message (Message, optional): Optional Message to filter by
+
+        Yields:
+            Metamessage: The next Metamessage object of the requested query
+
+        """
         page = 1
         page_size = 50
-        get_messages_response = self.get_messages(page, page_size)
+        get_metamessages_page = self.get_metamessages(metamessage_type=metamessage_type, message=message, page=page, page_size=page_size)
         while True:
             # get_session_response = self.get_sessions(user_id, location_id, page, page_size)
-            for message in get_messages_response.messages:
-                yield message
+            for metamessage in get_metamessages_page.items:
+                yield metamessage
 
-            new_messages = get_messages_response.next()
+            new_messages = get_metamessages_page.next()
             if not new_messages:
                 break
            
-            get_messages_response = new_messages
+            get_metamessages_page = new_messages
+
         
     def update(self, session_data: Dict):
-        """Update the metadata of a session
+        """Update the session_data of a session
 
         Args:
-            session_data (Dict): The Session object containing any new metadata
+            session_data (Dict): The Session object containing any new session_data
 
 
         Returns:
@@ -257,20 +461,10 @@ class Session:
         self.session_data = session_data
         return success
 
-    def delete(self):
-        """Delete a session by marking it as inactive"""
+    def close(self):
+        """Closes a session by marking it as inactive"""
         url = f"{self.common_prefix}/users/{self.user_id}/sessions/{self.id}"
         response = self.client.delete(url)
+        response.raise_for_status()
         self._is_active = False
 
-
-# class Message:
-#     def __init__(self, session: Session, id: int, is_user: bool, content: str):
-#         """Constructor for Message"""
-#         self.session = session
-#         self.id = id
-#         self.is_user = is_user
-#         self.content = content
-
-#     def __str__(self):
-#         return f"Message(id={self.id}, is_user={self.is_user}, content={self.content})"